refactor(common): move AbstractDistributedTransactionProvider from api to common
AbstractDistributedTransactionProvider is an internal abstract helper base, not part of the public
API; its sibling abstract bases (AbstractDistributedTransaction, AbstractDistributedTransactionManager,
AbstractTwoPhaseCommitTransaction, ...) already live in com.scalar.db.common, so it belongs there
too. The public SPI contract, the DistributedTransactionProvider interface, stays in com.scalar.db.api.

The in-tree subclasses (ConsensusCommitProvider, JdbcTransactionProvider) are updated to the new
import. Out-of-tree subclasses that extend this base (e.g. the cluster's transaction providers) must
update their import to com.scalar.db.common.AbstractDistributedTransactionProvider in a paired change.
